feat: implement skeleton of missions
This commit is the role management for the mission contract.
It supports roles for the sponsor, the contributor, the verifier and a
manager role. Initially the manager role will be owned by the sponsor.
This is laying the ground for the next versions.

There's a simple access application mechanism where individuals can
apply to become contributors for a mission.

The contract factory is responsible for configuring and deploying the
mission contracts.

As the protocol stewards we keep ownership of all the contracts so we
can fix things when they go wrong. This is only temporary until the
protocol has been built out completely.