feat: re-implemented module management
This patch is way too big, but the deadline for the hackathon is drawing
closer and closer.

This introduces a new way of handling modules. From tracking them to
being able to compose them. Authentication and eligibility is now
completely handled via eligibility modules as the interface is the same
and conceptually they're very close.

The eligibility module composability is very similar to Gnosis SAFE
modules.

Role management on the mossion contract is now inlined as the
MissionRoleManager was almost exclusively wrapping AccessControl methods
and just adding authentication on top. But since that's now handled via
the eligibility modules there's no need to have it.